% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_map.R
\name{hist_gadm}
\alias{hist_gadm}
\title{Create a list of historical map}
\usage{
hist_gadm(country, hash, lst_history, from = "1960", to = "2020",
  d.hash = NULL)
}
\arguments{
\item{country}{character string, name of the country to download}

\item{hash}{named character vector containing the translation in English
(standardized version) of the admin1 names. See `Details` for more
information.}

\item{lst_history}{A list containing a list of event, each code with a slot
after, a slot before, a slotevent (split/merge/rename/ complexe merge/
complexe split) and a slot year. See `Details` for more information.}

\item{from}{Initial date of the time range selected, of the class Date,
character or numeric. By default "1960".}

\item{to}{Final date of the time range selected, of the class Date, character
or numeric, by default "2020"}

\item{d.hash}{used in case of `complexe split` or `complexe merge` in the
`lst_history` object.  named character vector containing the translation in
English (standardized version) of the admin2 names. See `Details` for more
information.}
}
\value{
list of `sf` object
}
\description{
Create a list of historical map
}
\examples{

library(dictionary)

th_map <- hist_gadm("Thailand", th_province, th_history)

}
