% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_map.R
\name{hist_map}
\alias{hist_map}
\title{Create a list of historical map}
\usage{
hist_map(country, hash, lst_history, from = "1960", to = "2020",
  d.hash = NULL, tolerance = 0.01, path = NULL, file_rm = FALSE)
}
\arguments{
\item{country}{character string, name of the country to download.}

\item{hash}{named character vector containing the translation in English
(standardized version) of the admin1 names. See `Details` for more
information.}

\item{lst_history}{A list containing a list of event, each code with a slot
after, a slot before, a slotevent (split/merge/rename/ complexe merge/
complexe split) and a slot year. See `Details` for more information.}

\item{from}{Initial date of the time range selected, of the class Date,
character or numeric. By default "1960".}

\item{to}{Final date of the time range selected, of the class Date, character
or numeric, by default "2020".}

\item{d.hash}{used in case of `complexe split` or `complexe merge` in the
`lst_history` object.  named character vector containing the translation in
English (standardized version) of the admin2 names. See `Details` for more
information.}

\item{tolerance}{numeric for thinning (simplification). the tolerance value
should be in the metric of the input object (cf. from function
\code{\link[maptools]{thinnedSpatialPoly}}). By default, tolerance = 0.01.}

\item{path}{character string, name where the dowloaded file is saved.}

\item{file_rm}{boolean, if TRUE, remove the dowmloaded file.
By default, TRUE.}
}
\value{
list of `sf` object containing the maps of admin1 administrative
boundaries and two maps of the country boundaries (one in high resolution
and one in low resolution).
}
\description{
From a time range (by default: 1960-01-01 / 2020-12-31), recreates old map
by merging back together or spliting admin1 polygons from the current admin1
administrative boundaries downloaded from GADM \url{https://gadm.org}. Two
maps will be create for each year of event (split, merge or rename of
admin1), one in high resolution and one in low resolution.
}
\details{
The functions  needs a named vector, `hash` and `d.hash`
arguments, to translate the `NAME_1` column (and `NAME_2` if necessary) from
GADM \url{https://gadm.org} in a standardized English version. We advice to
use the named vector `xx_province` for admin1 or `xx_district` for admin2
contained in the `dictionary` package, for example:
\code{\link[dictionary]{kh_province}}. If no `hash` and/or `d.hash` arguments
is missing the column(s) `NAME_1` and/or `NAME_2` are encoded in UNICODE and
keep in native language.
\cr\cr
The function requires also a list of event (split/merge/rename/
complexe merge/complexe split) in a standardized format to recreate
historical map. We advice to use or the copy the format of the list
`xx_history` contained in the package `dictionary`.
For example: \code{\link[dictionary]{kh_history}}.
\cr\cr
The package `dictionary` is available on GitHub, to install it, it necessary
to have the `devtools` package:
`devtools::install_github("choisy/dictionary")`
\cr\cr
The function performs \code{\link[maptools]{thinnedSpatialPoly}} on
each map object with the tolerance (argument `tolerance`) value in the metric
of the input object.
}
\examples{
library(dictionary)

kh_map <- hist_map("Cambodia", kh_province, kh_history)

}
