% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{map_data}
\alias{map_data}
\title{Creates a list of historical map in a package}
\usage{
map_data(path, country, hash, lst_history, from = "1960", to = "2020",
  d.hash = NULL, tolerance = 0.01)
}
\arguments{
\item{path}{character string path of the package.}

\item{country}{character string, name of the country to download.}

\item{hash}{named character vector containing the translation in English
(standardized version) of the admin1 names. See \code{Details} for more
information.}

\item{lst_history}{A list containing a list of event, each code with a slot
after, a slot before, a slotevent (split/merge/rename/complex merge/
complex split) and a slot year. See \code{Details} for more information.}

\item{from}{Initial date of the time range selected, of the class Date,
character or numeric. By default "1960".}

\item{to}{Final date of the time range selected, of the class Date, character
or numeric, by default "2020".}

\item{d.hash}{used in case of \code{complex split} or \code{complex merge}
in the \code{lst_history} object.  named character vector containing the
translation in English (standardized version) of the admin2 names. See
\code{Details} for more information.}

\item{tolerance}{numeric for thinning (simplification). the tolerance value
should be in the metric of the input object (cf. from function
\code{\link[maptools]{thinnedSpatialPoly}}). By default, tolerance = 0.01.}
}
\description{
Creates a list of historical map with the function
\code{\link[histgadm]{hist_map}} and stores the download GADM
(\url{https://gadm.org}) files in a folder data_raw and the output maps in
the folder data of a package.
}
\examples{
library(dictionary)
\dontrun{
map_data("PACKAGE", "Cambodia", kh_province, kh_history)
}
}
