% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package.R
\name{map_documentation}
\alias{map_documentation}
\title{Makes documentation for sf map in a package}
\usage{
map_documentation(path)
}
\arguments{
\item{path}{character string path of the package.}
}
\description{
Create a Rd file for the data, containing the documentation for each sf
object in the package.
}
