% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_map.R
\name{hist_map}
\alias{hist_map}
\title{Create a list of historical map}
\usage{
hist_map(country, hash, lst_history, from = "1960", to = "2020",
  d.hash = NULL, tolerance = 0.01, save = FALSE, path = NULL,
  intlib = TRUE, lst_province_year = NULL)
}
\arguments{
\item{country}{character string, name of the country to download.}

\item{hash}{named character vector containing the translation in English
(standardized version) of the admin1 names. See `Details` for more
information.}

\item{lst_history}{A list containing a list of event, each code with a slot
after, a slot before, a slotevent (split/merge/rename/ complexe merge/
complexe split) and a slot year. See `Details` for more information.}

\item{from}{Initial date of the time range selected, of the class Date,
character or numeric. By default "1960".}

\item{to}{Final date of the time range selected, of the class Date, character
or numeric, by default "2020".}

\item{d.hash}{used in case of `complexe split` or `complexe merge` in the
`lst_history` object.  named character vector containing the translation in
English (standardized version) of the admin2 names. See `Details` for more
information.}

\item{tolerance}{numeric for thinning (simplification). the tolerance value
should be in the metric of the input object (cf. from function
\code{\link[maptools]{thinnedSpatialPoly}}). By default, tolerance = 0.01.}

\item{save}{boolean, specifies whether the downloaded file should be saved
in a specific path or not. If \code{NULL}, it will be asked interactively.
By default \code{FALSE}.}

\item{path}{character string, path to save the downloaded file. If
\code{NULL}, the file will be saved in the working directory. By default
\code{NULL}.}

\item{intlib}{boolean, specifies whether the downloaded file should be saved
in the library of packages. If \code{NULL}, it will be asked interactively.
By default \code{TRUE}.}

\item{lst_province_year}{A list containing the spatial expression of admin1
for each year of change, use to select the map expressed with the right
admin1 definition in time. See `Details` for more inforamtion}
}
\value{
a list of `sf` object containing the maps of admin1 administrative
boundaries and two maps of the country boundaries (one in high resolution
and one in low resolution).
}
\description{
From a time range (by default: 1960-01-01 / 2020-12-31), recreates old map
by merging back together or spliting admin1 polygons from the current admin1
administrative boundaries downloaded from GADM \url{https://gadm.org}. Two
maps will be create for each year of event (split, merge or rename of
admin1), one in high resolution and one in low resolution.
}
\details{
The functions  needs a named vector, `hash` and `d.hash`
arguments, to translate the `NAME_1` column (and `NAME_2` if necessary) from
GADM \url{https://gadm.org} in a standardized English version. We advice to
use the named vector `xx_province` for admin1 or `xx_district` for admin2
contained in the `dictionary` package, for example:
\code{\link[dictionary]{kh_province}}. If no `hash` and/or `d.hash` arguments
is missing the column(s) `NAME_1` and/or `NAME_2` are encoded in UNICODE and
keep in native language.
\cr\cr
The function needs also a list of event (split/merge/rename/
complexe merge/complexe split) in a standardized format to recreate
historical map. We advice to use or the copy the format of the list
`xx_history` contained in the package `dictionary`.
For example: \code{\link[dictionary]{kh_history}}.
If no list are inputed in the `lst_history` argument, only the current map of
admin1 administrative boundary and the country boundary in high and low
resolution in a list will be created.
\cr\cr
The package `dictionary` is available on GitHub, to install it, it necessary
to have the `devtools` package:
`devtools::install_github("choisy/dictionary")`
\cr\cr
The function performs \code{\link[maptools]{thinnedSpatialPoly}} on
each map object with the tolerance (argument `tolerance`) value in the metric
of the input object.
\cr\cr
The function uses the function \code{\link[sptools]{gadm}} from the package
\code{gadm}, to have more information on the parameters \code{save},
\code{path} and \code{intlib}, please take a look at the help of this
function.
\cr\cr
The arguments `lst_province_year` should be input as a list of charactor
vector containing the names of the admin1 written in a same way as `hash`
and/or `lst_history` ordered by year of change in administrative boundaries.
We advice to use or the copy the format of the list `xx_province_year`
contained in the package `dictionary`. For example:
\code{\link[dictionary]{kh_province_year}}.
\cr\cr
The output of the function is a named list: the admin1 boundaries named are
named as: the 2 characters ISO code, the year of expression of this admin1
administrative boundaries and the resolution. For example: "vn_1997_2004_high"
for the admin1 boundaries of Vietnam from 1997-01-01 until 2004-01-01 (not
include) in high quality.
}
\examples{
library(dictionary)

kh_map <- hist_map("Cambodia", kh_province, kh_history, path = FALSE)

}
