% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_package.R
\name{initial_pkg}
\alias{initial_pkg}
\title{Configures the intial files of the package}
\usage{
initial_pkg(path, name_pkg)
}
\arguments{
\item{path}{character string, path of the package.}

\item{name_pkg}{character string, name of the package.}
}
\description{
Creates a package for the gadm data, allow you to download (if wanted) data
from gadm and recreates historical map from a time range (interactive input)
}
