% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_data.R
\name{map_data}
\alias{map_data}
\title{Creates a list of historical map in a package}
\usage{
map_data(pckg_path, country, hash, lst_history, from = "1960",
  to = "2020", d.hash = NULL, intlib = FALSE, save = TRUE,
  force = FALSE, tolerance = NULL, lst_province_year = NULL,
  append_country = FALSE)
}
\arguments{
\item{pckg_path}{character string path of the package.}

\item{country}{character string, name of the country to download.}

\item{hash}{named character vector containing the translation in English
(standardized version) of the admin1 names. See \code{Details} for more
information.}

\item{lst_history}{A list containing a list of event, each code with a slot
after, a slot before, a slotevent (split/merge/rename/complex merge/
complex split) and a slot year. See \code{Details} for more information.}

\item{from}{Initial date of the time range selected, of the class Date,
character or numeric. By default "1960".}

\item{to}{Final date of the time range selected, of the class Date, character
or numeric, by default "2020".}

\item{d.hash}{used in case of \code{complex split} or \code{complex merge}
in the \code{lst_history} object.  named character vector containing the
translation in English (standardized version) of the admin2 names. See
\code{Details} for more information.}

\item{intlib}{boolean, specifies whether the downloaded file should be saved
in the library of packages. If \code{NULL}, it will be asked interactively.
By default \code{TRUE}.}

\item{save}{boolean, specifies whether the downloaded file should be saved
in a data-raw folder path or not. By default \code{TRUE}.
If \code{NULL}, it will be asked interactively.}

\item{force}{boolean, force to download the file even if already in the path.
By default \code{FALSE}.}

\item{tolerance}{numeric for thinning (simplification). the tolerance value
should be in the metric of the input object (cf. from function
\code{\link[maptools]{thinnedSpatialPoly}}). By default, tolerance = NULL.}

\item{lst_province_year}{A list containing the spatial expression of admin1
for each year of change, use to select the map expressed with the right
admin1 definition in time. See \code{Details} for more information.}

\item{append_country}{boolean, append the country level in the
final list. by default, FALSE}
}
\description{
Creates a list of historical map with the function
\code{\link[histgadm]{hist_map}} and stores the download GADM
(\url{https://gadm.org}) files in a folder data_raw and the output maps in
the folder data of a package.
}
\examples{
library(dictionary)
\dontrun{
map_data("PACKAGE", "Cambodia", kh_province, kh_history)
}
}
