% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_documentation.R
\name{map_documentation}
\alias{map_documentation}
\title{Makes documentation for sf map in a package}
\usage{
map_documentation(path)
}
\arguments{
\item{path}{character string path of the package.}
}
\description{
Create a \code{.R} file containing the roxygen comment and \code{.Rd} files
corresponding, containing the documentation or each sf object in a folder
 \code{data} and \code{man} of a package.
If a \code{data.R} file already exists in the folder \code{R} of a package,
it will be overwritten.
}
